requirements.txt
httpx==0.27.2
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
yfinance==0.2.50
pydantic==2.9.2
python-dateutil==2.9.0.post0
pytest==8.3.3
